autorun: return None if there was no setting for autorun enable
This should allow caller to possibly handle state when no configuration
is set. No other code should be dirrectly affected because None
evaluates to False in boolean so we are able to still use this function
the same way in conditions. The default value is still False but
representation is None now.