svupdater/lists: add support for labels
Labels are new generic approach to existing official boolean option.
This removes that option and instead introduces flags that can symbol
anything.

This implementation is just extension of existing one so everything is
till returned all at once. Labels can be matched by name between various
pkglists but that has to be done by user. Here info for labels is
duplicated to every occurrence of that label.