Add support for boards package lists filter
This is filter for both package lists as well its options to skip them
if they are not compatible with current board. This is primarily
intended to support hardware specific extension in package lists.

No change is required outside of updater-supervisor as package lists are
just skipped and considered as not defined (not existing).