svupdater: do not run pkgupdate with --task-log
This removes task log. I was used as kind of history of changes but was
primarilly to observe long term execution. There are issues with it as
that we do not manage that log correctly (it is not logrotated). It is
also questionable how usable it is. One execution can be and is parsed
from state-log. This makes this file almost without use and not needed
for functionality at all. In the end making it adept for removal.