Add msgtrace module
This module is intended as "hack" to allow some state reporting of
updater to other components. It is really hack as it relies on debug
messages send to syslog.
This implementation does not match every single message produced by
updater. It only covers standard execution messages and only those with
some sort of "benefit" to user.