Add --autorun option and allow run even if autorun is disabled
The original idea was to use autorun configuration to disable automatic
updates. As it was implemented until now it was preventing execution of
updates not just automatic but manual as well. The only option was to
use pkgupdate directly. This should allow usage of supervisor not just
when automatic updates are enabled.