Fix check for installed packages and removed package
Updater preserves removed packages in index if they at time of removal
had some modified configuration files. Such packages are removed and
should be reported as such. There is also no control file on filesystem.

This fixes missing control file error by just not reading control file
if it is not present. This has no impact as info is not available about
that package and should not be required anyway.

This also fixes installed method by properly checking status fields.