Replace ping with curl for network check
In reality we need to download files from repo.turris.cz so it makes
sense to try perform real download rather than just ping.

This also changes when we check for network. We check every time that is
enabled. It makes sense as that way we can cover some temporally service
dropouts even when not running from cron.

There are other tiny tweaks such as that we no longer use name of
package in package itself and rather import relatively.