files/immediate_reboot: fix block on package removal
This makes sure that init script is not executed when we are installing
or removing package. The important problem this covers is the block of
execution on package removal as immediate_reboot init script waits for
termination of opkg or updater while at the same time opkg waits for
postinstall script termination and thus termination of the initial wait.
This creates deadlock.