notify: Do not send stacktrace for handled errors
Currently the supervisor sends out the stacktrace whenever it encounters
any known error. Much better outcome would be to send the actual error,
same as the output in console. Especially if we know that updater
handled the error. Typically the missing package.

This is PoC fix right now, could use some smoothing of the rough edges -
better formatting.